stremlining startup
Former-commit-id: 5691a9775d5925d388bbb606ef62937333f0286f

+++ dockers/machine/files/start.sh
@@ -1,5 +1,11 @@
 cd ${PROJECT_ROOT}/Gp/machine
-for metadata in `find ${PROJECT_ROOT}/Gp/machine/datasets/ | grep metadata`;do rm -f $metadata;done
-rm -f ${PROJECT_ROOT}/Gp/machine/specs.json
-pm2 start machine.config.js --watch
+if [ -f '/root/forum' ]; then
+    echo "forum exists."
+else
+    #clear out metadata and start fro scratch
+    rm -f ${PROJECT_ROOT}/Gp/machine/specs.json
+    for metadata in `find ${PROJECT_ROOT}/Gp/machine/datasets/ | grep metadata`;do rm -f $metadata;done
+    touch /root/forum
+fi
+    pm2 start machine.config.js
 bash
